Move lock component service to KM
