+++ apps/browser/src/key-management/lock/services/extension-lock-component.service.ts
@@ -4,11 +4,6 @@ import { inject } from "@angular/core";
 import { combineLatest, defer, map, Observable } from "rxjs";
 
 import {
-  BiometricsDisableReason,
-  LockComponentService,
-  UnlockOptions,
-} from "@bitwarden/auth/angular";
-import {
   PinServiceAbstraction,
   UserDecryptionOptionsServiceAbstraction,
 } from "@bitwarden/auth/common";
@@ -17,9 +12,14 @@ import { PlatformUtilsService } from "@bitwarden/common/platform/abstractions/pl
 import { KeySuffixOptions } from "@bitwarden/common/platform/enums";
 import { UserId } from "@bitwarden/common/types/guid";
 import { KeyService, BiometricsService } from "@bitwarden/key-management";
+import {
+  LockComponentService,
+  BiometricsDisableReason,
+  UnlockOptions,
+} from "@bitwarden/key-management/angular";
 
-import { BiometricErrors, BiometricErrorTypes } from "../models/biometricErrors";
-import { BrowserRouterService } from "../platform/popup/services/browser-router.service";
+import { BiometricErrors, BiometricErrorTypes } from "../../../models/biometricErrors";
+import { BrowserRouterService } from "../../../platform/popup/services/browser-router.service";
 
 export class ExtensionLockComponentService implements LockComponentService {
   private readonly userDecryptionOptionsService = inject(UserDecryptionOptionsServiceAbstraction);
