+++ apps/web/src/app/auth/core/services/index.ts
@@ -3,4 +3,3 @@ export * from "./login-decryption-options";
 export * from "./webauthn-login";
 export * from "./set-password-jit";
 export * from "./registration";
-export * from "./web-lock-component.service";
