+++ apps/web/src/app/key-management/lock/services/web-lock-component.service.spec.ts
